3-Multi-Container-Application/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2==2.9.3
Flask-Migrate